Extract definition validation to validation file
Cut-paste, no code changes
